Migrate most of the tests to the Rust MeTTa runner
Keep Python MeTTa runner for the tests which are not compatible yet.
Move common.py to common_old.py. Rename MeTTa to MeTTaPy. Implement
new MeTTa which uses Rust runner.
