+++ python/tests/test_run_metta.py
@@ -1,6 +1,6 @@
 import unittest
 
-from common import MeTTa
+from common import MeTTa, MeTTaPy
 
 class MeTTaTest(unittest.TestCase):
 
@@ -86,6 +86,6 @@ class MeTTaTest(unittest.TestCase):
         self.process_exceptions(MeTTa().import_file("scripts/d2_higherfunc.metta"))
         self.process_exceptions(MeTTa().import_file("scripts/d3_deptypes.metta"))
         self.process_exceptions(MeTTa().import_file("scripts/d4_type_prop.metta"))
-        self.process_exceptions(MeTTa().import_file("scripts/d5_auto_types.metta"))
-        self.process_exceptions(MeTTa().import_file("scripts/e1_kb_write.metta"))
+        self.process_exceptions(MeTTaPy().import_file("scripts/d5_auto_types.metta"))
+        self.process_exceptions(MeTTaPy().import_file("scripts/e1_kb_write.metta"))
